Move get_filters and add support for Py3 dicts
